refactor(jpush): fix load file error
