--- app/util/jpush.ts
@@ -1,29 +1,0 @@
-'use strict';
-const env = process.env;
-/**
- * jpush util
- */
-
-const JPush = require('jpush-sdk');
-const jpushClient = JPush.buildClient(
-  env.JPUSH_APP_kEY,
-  env.JPUSH_MASTER_SECRET,
-);
-export function pushMessage(alias, alert, title) {
-  jpushClient
-    .push()
-    .setPlatform(JPush.ALL)
-    .setAudience(JPush.alias(alias))
-    .setNotification(
-      JPush.android(alert, title, 1),
-      JPush.ios(alert, 'sound', 1, null, { msg: title })
-    )
-    .send((err, res) => {
-      if (err) {
-        console.log(err.message);
-      } else {
-        console.log('Sendno: ' + res.sendno);
-        console.log('Msg_id: ' + res.msg_id);
-      }
-    });
-}
